Safari-609.1.11.1/LayoutTests/fast/forms/search/intrinsic-search-width-with-decoration-border-padding.html

<!DOCTYPE html>
<style>
#search1::-webkit-search-cancel-button {
    border: 10px solid lime;
    padding: 10px;
}
#search1::-webkit-search-decoration {
    border: 10px solid lime;
    padding: 10px;
}
</style>
<body>
<script src="../../../resources/js-test-pre.js"></script>

<input type="search" id="search1">
<br/>
<input type="search" id="search2">
<script>
shouldBe('document.getElementById("search1").offsetWidth', 'document.getElementById("search2").offsetWidth');
</script>

<script src="../../../resources/js-test-post.js"></script>
</body>
