LayoutTests/platform/ios-sim/fast/forms/search/intrinsic-search-width-with-decoration-border-padding-expected.txt

FAIL document.getElementById("search1").offsetWidth should be 136. Was 135.
PASS successfullyParsed is true

TEST COMPLETE
 
